refactor: InterviewService 인터페이스 분리 #75
